requirements.txt
jupyter==1.0.0
requests==2.31.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
PyPDF2==3.0.1
PyAutoGUI==0.9.54
beautifulsoup4==4.12.3
pillow==10.2.0
scikit-learn==1.4.0
folium==0.15.1
pycoingecko==3.1.0
opencv-python==4.9.0.80
pytesseract==0.3.10
nltk==3.8.1
Mesa==2.2.4
Django==5.0.2
tensorflow==2.15.0
dash==2.15.0
Flask==3.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
flickrapi==2.4.0
xlwt==1.3.0
openpyxl==3.1.2
selenium==4.17.2
webdriver-manager==4.0.1
pykafka==2.8.0
yfinance==0.2.36
ta==0.11.0
openrouteservice==2.3.3
statsmodels==0.14.1
pygame==2.5.2
PuLP==2.8.0
mip==1.15.0
boto3==1.34.39
keras==2.15.0
pathlib==1.0.1
tabulate==0.9.0
autopep8==2.0.4
pycodestyle==2.11.1
pylint==3.0.3
pyarrow==15.0.0
prettytable==3.9.0
geopandas==0.14.3
openai
streamlit
PyPDF2
plotly
openpyxl
xlsxwriter
pymysql
flask
mysql-connector-python
schedule
selenium
selenium webdriver manager
geopy
contextily
docker
Pillow
torch